% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageName-package.R
\docType{package}
\name{PackageName-package}
\alias{PackageName}
\alias{PackageName-package}
\title{PackageName: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LandSciTech/PackageName}
  \item Report bugs at \url{https://github.com/LandSciTech/PackageName/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{email@email.ca} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

Other contributors:
\itemize{
  \item His Majesty the King in Right of Canada, as represented by the Minister of Environment and Climate Change [copyright holder]
}

}
\keyword{internal}
